scraper/requirements.txt
playwright
httpx
python-dotenv
pytest
pylint